extract to a class
